+++ library/src/main/java/com/liulishuo/filedownloader/services/IFileDownloadDBHelper.java
@@ -44,7 +44,7 @@ interface IFileDownloadDBHelper {
 
     void remove(final int id);
 
-    void update(final int id, final int status, final int soFar, final int total);
+    void update(final int id, final byte status, final int soFar, final int total);
 
     void updateHeader(final int id, final String etag);
